Revert add, edit claimset request and validator file update
