+++ Application/EdFi.Ods.AdminApi/Features/ClaimSets/EditClaimSet.cs
@@ -6,9 +6,11 @@
 using AutoMapper;
 using EdFi.Ods.AdminApi.Infrastructure;
 using EdFi.Ods.AdminApi.Infrastructure.ClaimSetEditor;
+using EdFi.Ods.AdminApi.Infrastructure.Database.Queries;
 using EdFi.Ods.AdminApi.Infrastructure.ErrorHandling;
 using FluentValidation;
 using FluentValidation.Results;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace EdFi.Ods.AdminApi.Features.ClaimSets;
 
@@ -22,7 +24,7 @@ public class EditClaimSet : IFeature
         .BuildForVersions(AdminApiVersions.V2);
     }
 
-    public async Task<IResult> Handle(EditClaimSetValidator validator, IEditClaimSetCommand editClaimSetCommand,
+    public async Task<IResult> Handle(Validator validator, IEditClaimSetCommand editClaimSetCommand,
         UpdateResourcesOnClaimSetCommand updateResourcesOnClaimSetCommand,
         IGetClaimSetByIdQuery getClaimSetByIdQuery,
         IGetResourcesByClaimSetIdQuery getResourcesByClaimSetIdQuery,
@@ -68,6 +70,93 @@ public class EditClaimSet : IFeature
         return Results.Ok(model);
     }
 
+    [SwaggerSchema(Title = "EditClaimSetRequest")]
+    public class EditClaimSetRequest
+    {
+        [SwaggerSchema(Description = "ClaimSet id", Nullable = false)]
+        public int Id { get; set; }
 
-  
+        [SwaggerSchema(Description = FeatureConstants.ClaimSetNameDescription, Nullable = false)]
+        public string? Name { get; set; }
+
+        [SwaggerSchema(Description = FeatureConstants.ResourceClaimsDescription, Nullable = false)]
+        public List<ResourceClaimModel>? ResourceClaims { get; set; }
+    }
+
+    public class Validator : AbstractValidator<EditClaimSetRequest>
+    {
+        private readonly IGetClaimSetByIdQuery _getClaimSetByIdQuery;
+        private readonly IGetAllClaimSetsQuery _getAllClaimSetsQuery;
+
+        public Validator(IGetClaimSetByIdQuery getClaimSetByIdQuery,
+            IGetAllClaimSetsQuery getAllClaimSetsQuery,
+            IGetResourceClaimsAsFlatListQuery getResourceClaimsAsFlatListQuery,
+            IGetAllAuthorizationStrategiesQuery getAllAuthorizationStrategiesQuery)
+        {
+            _getClaimSetByIdQuery = getClaimSetByIdQuery;
+            _getAllClaimSetsQuery = getAllClaimSetsQuery;
+
+            var resourceClaims = (Lookup<string, ResourceClaim>)getResourceClaimsAsFlatListQuery.Execute()
+                .ToLookup(rc => rc.Name?.ToLower());
+
+            var authStrategyNames = getAllAuthorizationStrategiesQuery.Execute()
+                .Select(a => a.AuthStrategyName).ToList();
+
+            RuleFor(m => m.Id).NotEmpty();
+
+            RuleFor(m => m.Id)
+                .Must(BeAnExistingClaimSet)
+                .WithMessage(FeatureConstants.ClaimSetNotFound);
+
+            RuleFor(m => m.Name)
+            .NotEmpty()
+            .Must(BeAUniqueName)
+            .WithMessage(FeatureConstants.ClaimSetAlreadyExistsMessage)
+            .When(m => BeAnExistingClaimSet(m.Id) && NameIsChanged(m));
+
+            RuleFor(m => m.Name)
+                .MaximumLength(255)
+                .WithMessage(FeatureConstants.ClaimSetNameMaxLengthMessage);
+
+            RuleFor(m => m).Custom((claimSet, context) =>
+            {
+                var resourceClaimValidator = new ResourceClaimValidator();
+
+                if (claimSet.ResourceClaims != null && claimSet.ResourceClaims.Any())
+                {
+                    foreach (var resourceClaim in claimSet.ResourceClaims)
+                    {
+                        resourceClaimValidator.Validate(resourceClaims, authStrategyNames,
+                            resourceClaim, claimSet.ResourceClaims, context, claimSet.Name);
+                    }
+                }
+            });
+        }
+
+        private bool BeAnExistingClaimSet(int id)
+        {
+            try
+            {
+                _getClaimSetByIdQuery.Execute(id);
+                return true;
+            }
+            catch (AdminApiException)
+            {
+                throw new NotFoundException<int>("claimSet", id);
+            }
+        }
+
+        private bool NameIsChanged(EditClaimSetRequest model)
+        {
+            return _getClaimSetByIdQuery.Execute(model.Id).Name != model.Name;
+        }
+
+        private bool BeAUniqueName(string? name)
+        {
+            return _getAllClaimSetsQuery.Execute().All(x => x.Name != name);
+        }
+    }
+
+
+
 }
